implemented deleteReview so it deletes both copies of reviews

+++ skeleton/src/components/DeleteReviewForm.js
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from 'react';
-import { doc, deleteDoc } from 'firebase/firestore';
+import { doc, deleteDoc, getDoc } from 'firebase/firestore';
 import { db } from '../config/firebase'; 
 import Swal from 'sweetalert2';
 
@@ -7,19 +7,28 @@ function DeleteReview({ reviewIdToDelete, userId, onUpdate}) { // Corrected para
   const handleDelete = async () => {
     try {
       const userReviewsRef = doc(db, `users/${userId}/userReviews/${reviewIdToDelete}`);
-      await deleteDoc(userReviewsRef);
-
+      const userReviewsDoc = await getDoc(userReviewsRef);
+  
+      if (!userReviewsDoc.exists()) {
+        throw new Error("User review document doesn't exist");
+      }
+  
+      const userDocData = userReviewsDoc.data();
+      const otherReviewDoc = doc(db, "reviews", userDocData.reviewId);
+  
+      await Promise.all([
+        deleteDoc(userReviewsRef),
+        deleteDoc(otherReviewDoc)
+      ]);
+  
       Swal.fire({
         icon: 'success',
         title: 'Deleted!',
         showConfirmButton: false,
         timer: 1500,
       });
-
+  
       onUpdate(reviewIdToDelete); // Notify parent component to update
-
-
-      // Optionally update state or trigger a callback to handle updates
     } catch (error) {
       console.error('Error deleting review:', error);
       Swal.fire({
